+++ gems/aws-sdk-accessanalyzer/lib/aws-sdk-accessanalyzer/client.rb
@@ -3068,7 +3068,7 @@ module Aws::AccessAnalyzer
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-accessanalyzer'
-      context[:gem_version] = '1.90.0'
+      context[:gem_version] = '1.91.0'
       Seahorse::Client::Request.new(handlers, context)
     end