+++ gems/aws-sdk-connect/lib/aws-sdk-connect/client.rb
@@ -29090,7 +29090,7 @@ module Aws::Connect
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-connect'
-      context[:gem_version] = '1.256.0'
+      context[:gem_version] = '1.257.0'
       Seahorse::Client::Request.new(handlers, context)
     end