+++ gems/aws-sdk-databasemigrationservice/lib/aws-sdk-databasemigrationservice/client.rb
@@ -13723,7 +13723,7 @@ module Aws::DatabaseMigrationService
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-databasemigrationservice'
-      context[:gem_version] = '1.144.0'
+      context[:gem_version] = '1.145.0'
       Seahorse::Client::Request.new(handlers, context)
     end