+++ gems/aws-sdk-datazone/lib/aws-sdk-datazone/client.rb
@@ -16217,7 +16217,7 @@ module Aws::DataZone
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-datazone'
-      context[:gem_version] = '1.77.0'
+      context[:gem_version] = '1.78.0'
       Seahorse::Client::Request.new(handlers, context)
     end