+++ gems/aws-sdk-ec2/lib/aws-sdk-ec2/client.rb
@@ -73770,7 +73770,7 @@ module Aws::EC2
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-ec2'
-      context[:gem_version] = '1.617.0'
+      context[:gem_version] = '1.618.0'
       Seahorse::Client::Request.new(handlers, context)
     end