+++ gems/aws-sdk-ecs/lib/aws-sdk-ecs.rb
@@ -55,7 +55,7 @@ module Aws::ECS
   autoload :EndpointProvider, 'aws-sdk-ecs/endpoint_provider'
   autoload :Endpoints, 'aws-sdk-ecs/endpoints'
 
-  GEM_VERSION = '1.231.0'
+  GEM_VERSION = '1.232.0'
 
 end