+++ gems/aws-sdk-ecs/lib/aws-sdk-ecs/client.rb
@@ -16004,7 +16004,7 @@ module Aws::ECS
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-ecs'
-      context[:gem_version] = '1.231.0'
+      context[:gem_version] = '1.232.0'
       Seahorse::Client::Request.new(handlers, context)
     end