+++ gems/aws-sdk-glue/lib/aws-sdk-glue/client.rb
@@ -19063,7 +19063,7 @@ module Aws::Glue
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-glue'
-      context[:gem_version] = '1.257.0'
+      context[:gem_version] = '1.258.0'
       Seahorse::Client::Request.new(handlers, context)
     end