+++ gems/aws-sdk-mediapackagev2/lib/aws-sdk-mediapackagev2/client.rb
@@ -5322,7 +5322,7 @@ module Aws::MediaPackageV2
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-mediapackagev2'
-      context[:gem_version] = '1.63.0'
+      context[:gem_version] = '1.64.0'
       Seahorse::Client::Request.new(handlers, context)
     end