+++ gems/aws-sdk-mgn/lib/aws-sdk-mgn/client.rb
@@ -7651,7 +7651,7 @@ module Aws::Mgn
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-mgn'
-      context[:gem_version] = '1.67.0'
+      context[:gem_version] = '1.68.0'
       Seahorse::Client::Request.new(handlers, context)
     end