+++ gems/aws-sdk-mwaa/lib/aws-sdk-mwaa/client.rb
@@ -1580,7 +1580,7 @@ module Aws::MWAA
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-mwaa'
-      context[:gem_version] = '1.73.0'
+      context[:gem_version] = '1.74.0'
       Seahorse::Client::Request.new(handlers, context)
     end