+++ gems/aws-sdk-partnercentralselling/lib/aws-sdk-partnercentralselling/client.rb
@@ -4215,7 +4215,7 @@ module Aws::PartnerCentralSelling
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-partnercentralselling'
-      context[:gem_version] = '1.31.0'
+      context[:gem_version] = '1.32.0'
       Seahorse::Client::Request.new(handlers, context)
     end