+++ gems/aws-sdk-qconnect/lib/aws-sdk-qconnect/client.rb
@@ -9529,7 +9529,7 @@ module Aws::QConnect
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-qconnect'
-      context[:gem_version] = '1.55.0'
+      context[:gem_version] = '1.56.0'
       Seahorse::Client::Request.new(handlers, context)
     end