+++ gems/aws-sdk-quicksight/lib/aws-sdk-quicksight/client.rb
@@ -22268,7 +22268,7 @@ module Aws::QuickSight
         tracer: tracer
       )
       context[:gem_name] = 'aws-sdk-quicksight'
-      context[:gem_version] = '1.181.0'
+      context[:gem_version] = '1.182.0'
       Seahorse::Client::Request.new(handlers, context)
     end