+++ gems/aws-sdk-managedgrafana/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.63.0 (2026-05-19)
+------------------
+
+* Feature - Introduce degraded workspace status as a possible Amazon Managed Grafana workspace status, and a new field named degraded workspace reason which informs customers why the workspace is degraded in the DescribeWorkspace API response.
+
 1.62.0 (2026-05-14)
 ------------------